requirements.txt
Django==1.4.3
MySQL-python==1.2.4
PIL==1.1.7
Pillow==1.7.8
South==0.7.6
URLObject==2.3.1
Unidecode==0.04.12
argparse==1.2.1
distribute==0.6.34
django-ajax-selects==1.2.5
django-crispy-forms==1.2.3
django-dajax==0.9.2
django-dajaxice==0.5.5
django-debug-toolbar==0.9.4
django-less==0.7
django-pagination==1.0.7
django-photologue==2.5
django-recaptcha==0.0.6
django-tagging==0.3.1
djangorestframework==0.3.3
simplejson==3.0.7
wsgiref==0.1.2
